git-svn-id: http://ikube.googlecode.com/svn/trunk@1238 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/IConstants.java
@@ -144,7 +144,7 @@ public interface IConstants {
 	String STRIP_CHARACTERS = "|!,[]{};:/\\.-_";
 
 	long MAX_ACTIONS = 100000;
-	double MAX_SNAPSHOTS = 90;
+	long MAX_SNAPSHOTS = 90;
 	String TIER = "tier";
 
 	String SEARCH = "search";
